WIP: Bump irloader to C++.
Since we are in C++ now, ideally we don't need to convert LLVM IR to
json to translate it into Coq. But this is too involved. So we decide
to keep the original workflow as a initial step.

`irloader` is compiled as a shared library so hopefully the implementation
can be easily switched later on.

`json2coq` is compiled as another executable for testing.